services/reddit_discovery_agent/requirements.txt
openai
duckduckgo-search
python-dotenv
rich
fastapi
uvicorn
pydantic